fix some links (janky)
added css and js links to issue layout instead of include. Still in include for all files, and still gets missing file errors in console in issues, but this way links are loaded. also, added onerror to fix the src of the image in navbar for issue pages. also hardcoded the favicon link in issue layout